requirements.txt
blinker==1.6.2
click==8.1.3
dnspython==2.3.0
Faker==18.11.2
Flask==2.3.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
MongoFrames==1.3.6
passlib==1.7.4
pycairo==1.24.0
pymongo==4.4.0
python-dateutil==2.8.2
python-dotenv==1.0.0
six==1.16.0
Werkzeug==2.3.6
